+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/beneficiary/BeneficiaryDetail.kt
@@ -79,7 +79,12 @@ fun BeneficiaryDetail(
                         message = "Beneficiario eliminado correctamente"
                     )
                 }
-            }
+            },
+
+            // --- Arreglar el error ---
+            confirmText = "Confirmar",
+            dismissText = "Cancelar"
+            // --- Fin del arreglo :)
         )
     }