+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [0.1.0] - 10/11/2025
 
+### Fixed
+ - BeneficiaryDetail.kt fixed error that prevented merge
+
 ### Added
  - Added basic logic for US BEN-01 and BEN-02
  - Added DecisionDialog Atom